renaming the buttons to be more generic and repackaging
